+++ waspc/cli/src/Wasp/Cli/Command/Db/Studio.hs
@@ -13,12 +13,12 @@ import Wasp.Cli.Command.Require (InWaspProject (InWaspProject), require)
 import Wasp.Generator.DbGenerator.Jobs (runStudio)
 import Wasp.Job.IO (readJobMessagesAndPrintThemPrefixed)
 import qualified Wasp.Message as Msg
-import Wasp.Project.Common (generatedCodeDirInWaspProjectDir)
+import Wasp.Project.Common (generatedAppDirInWaspProjectDir)
 
 studio :: Command ()
 studio = do
   InWaspProject waspProjectDir <- require
-  let genProjectDir = waspProjectDir </> generatedCodeDirInWaspProjectDir
+  let genProjectDir = waspProjectDir </> generatedAppDirInWaspProjectDir
 
   cliSendMessageC $ Msg.Start "Running studio..."